update for latest rustc
I've tried my best to update this crate to a recent nightly
(`rustc1.13.0-nightly (d4037fc47 2016-09-16)`). Since this is a
pretty large patch and I'm unfamiliar with the codebase, there might be some
safety issues.

The examples run fine in release mode, but not in debug mode because large
arrays make the stack overflow. `Vec` doesn't implement BoehmTraced, so there is
no easy fix as far as I can tell.